Move panel-header code to layout/panels directory
